requirements.txt
Django==1.4
South==0.7.6
dj-database-url==0.2.1
honcho==0.2.0
django-social-auth==0.7.22
django-storages==1.1.5
boto==2.6.0
pyfaceb==0.2.3
psycopg2==2.4.6
gunicorn==0.17.2
requests==1.2.0
pyephem==3.7.5.1
python-dateutil==2.1
